# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dokandar Mama shop assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: withDueOnly
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Customer" }
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /customers/{id}/ledger:
    get:
      operationId: getCustomerLedger
      tags: [customers]
      summary: Get a customer's baki ledger entries
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LedgerEntry" }
  /customers/{id}/payments:
    post:
      operationId: recordCustomerPayment
      tags: [customers]
      summary: Record a baki payment from a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentInput" }
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
...
  /master-products:
    get:
      operationId: searchMasterProducts
      tags: [master-products]
      summary: Search the global product catalogue
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Matches
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MasterProduct" }
  /master-products/lookup:
    get:
      operationId: lookupMasterProduct
      tags: [master-products]
      summary: Barcode lookup in the global catalogue
      parameters:
        - name: barcode
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MasterProduct" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /cashbox/open:
    post:
      operationId: openCashbox
      tags: [cashbox]
      summary: Open the drawer with an opening balance
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CashboxOpenInput" }
      responses:
        "201":
          description: Opened
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CashSession" }
  /cashbox/movements:
    post:
      operationId: addCashMovement
      tags: [cashbox]
      summary: Record an expense or manual cash movement
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CashMovementInput" }
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CashMovement" }
  /cashbox/close:
    post:
      operationId: closeCashbox
      tags: [cashbox]
      summary: Close the drawer with counted cash
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CashboxCloseInput" }
      responses:
        "200":
          description: Closed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CashSession" }
  /cashbox/sessions:
    get:
      operationId: listCashSessions
      tags: [cashbox]
      summary: Cash session history
      parameters:
        - name: from
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CashSession" }
...
components:
  schemas:
...
    MasterProduct:
      type: object
      required: [id, barcode, name]
      properties:
        id: { type: integer }
        barcode: { type: string }
        name: { type: string }
        nameBn: { type: ["string", "null"] }
        brand: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        unit: { type: ["string", "null"] }
        defaultPrice: { type: ["number", "null"] }
        imageUrl: { type: ["string", "null"] }
    CashSession:
      type: object
      required: [id, status, openingBalance, openedAt]
      properties:
        id: { type: integer }
        status: { type: string, enum: [open, closed] }
        openingBalance: { type: number }
        expectedClosing: { type: ["number", "null"] }
        countedClosing: { type: ["number", "null"] }
        difference: { type: ["number", "null"] }
        note: { type: ["string", "null"] }
        openedAt: { type: string }
        closedAt: { type: ["string", "null"] }
    CashMovement:
      type: object
      required: [id, type, amount, createdAt]
      properties:
        id: { type: integer }
        type: { type: string, enum: [sale, expense, cash_in, cash_out] }
        amount: { type: number }
        note: { type: ["string", "null"] }
        saleId: { type: ["integer", "null"] }
        createdAt: { type: string }
...
    CashboxOpenInput:
      type: object
      required: [openingBalance]
      properties:
        openingBalance: { type: number, minimum: 0 }
        note: { type: string }
    CashMovementInput:
      type: object
      required: [type, amount]
      properties:
        type: { type: string, enum: [expense, cash_in, cash_out] }
        amount: { type: number, minimum: 0 }
        note: { type: string }
    CashboxCloseInput:
      type: object
      required: [countedClosing]
      properties:
        countedClosing: { type: number, minimum: 0 }
        note: { type: string }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Customer:
      type: object
      required: [id, name, bakiBalance, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: ["string", "null"] }
        bakiBalance: { type: number }
        phoneVerified: { type: boolean }
        verifiedAt: { type: ["string", "null"] }
        createdAt: { type: string }
    CustomerInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string }
    CustomerUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string }
    LedgerEntry:
      type: object
      required: [id, customerId, type, amount, balanceAfter, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        type: { type: string, enum: [sale, payment] }
        amount: { type: number }
        balanceAfter: { type: number }
        note: { type: ["string", "null"] }
        createdAt: { type: string }
    PaymentInput:
      type: object
      required: [amount]
      properties:
        amount: { type: number, exclusiveMinimum: 0 }
        note: { type: string }
...