# lib/api-zod/src/generated/types/cashboxCloseInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Dokandar Mama shop assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CashboxCloseInput {
  /** @minimum 0 */
  countedClosing: number;
  note?: string;
}