# lib/api-zod/src/generated/types/masterProduct.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Dokandar Mama shop assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface MasterProduct {
  id: number;
  barcode: string;
  name: string;
  /** @nullable */
  nameBn?: string | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  defaultPrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
}